Refactoring of track management
